requirements.txt
torch
torchaudio
pandas
numpy
scikit-learn
tqdm
pyyaml
soundfile
scipy
transformers
onnxruntime
hydra-core
inflect
jiwer
inflect
jiwer
pystoi
